make transient/deferredrelease nilling part of side effects
Nilling of objects runs deinits which can possibly trigger new actions.
Before old side effects could release the mutex and switch to main thread
Then a queued action the mutex, and the first one blocks on the main thread
waiting for mutex.
Then the second one runs and gets to nilling of transient objects which tries
to run stuff on main thread and blocks and we have a deadlock.

Somehow this broke one test. It was running the action and immediately
I couldn't figure out what it was testing. So I deleted it.